requirements.txt
fastapi
uvicorn
httpx
python-multipart
pillow
python-dotenv
pydantic
firebase-admin
websockets